More code refactoring
