+++ src/Nodes/Create.ts
@@ -1,0 +1,199 @@
+import { CmdSyntaxKind, NodeFlag } from "./Enum";
+import {
+	InterpolatedStringExpression,
+	StringLiteral,
+	CommandName,
+	Node,
+	CommandStatement,
+	InnerExpression,
+	PrefixToken,
+	PrefixExpression,
+	CommandSource,
+	NumberLiteral,
+	Identifier,
+	Option,
+	OperatorToken,
+	VariableDeclaration,
+	VariableStatement,
+	BooleanLiteral,
+	EndOfStatement,
+	InvalidNode,
+	BinaryExpression,
+	NodeError,
+} from "./NodeTypes";
+import { isNode } from "./Guards";
+
+export function createInterpolatedString(
+	...values: InterpolatedStringExpression["values"]
+): InterpolatedStringExpression {
+	const expression: InterpolatedStringExpression = { kind: CmdSyntaxKind.InterpolatedString, values, flags: 0 };
+	for (const value of values) {
+		value.parent = expression;
+	}
+	return expression;
+}
+
+export function createNodeError(message: string, node: Node): NodeError {
+	return {
+		node,
+		message,
+	};
+}
+
+/**
+ * Flattens an interpolated string into a regular string
+ * @param expression The interpolated string expression
+ * @param variables The variables for identifiers etc. to flatten with
+ */
+export function flattenInterpolatedString(
+	expression: InterpolatedStringExpression,
+	variables: Record<string, defined>,
+): StringLiteral {
+	let text = "";
+	for (const value of expression.values) {
+		if (isNode(value, CmdSyntaxKind.Identifier)) {
+			text += tostring(variables[value.name]);
+		} else {
+			text += value.text;
+		}
+	}
+	return { text, kind: CmdSyntaxKind.String, flags: 0 };
+}
+
+export function createCommandStatement(command: CommandName, children: Node[], startPos?: number, endPos?: number) {
+	const statement: CommandStatement = {
+		kind: CmdSyntaxKind.CommandStatement,
+		command,
+		children,
+		flags: 0,
+		pos: startPos,
+		endPos,
+	};
+	for (const child of statement.children) {
+		child.parent = statement;
+	}
+
+	return statement;
+}
+
+export function createInnerExpression(expression: InnerExpression["expression"], startPos?: number, endPos?: number) {
+	const statement: InnerExpression = {
+		kind: CmdSyntaxKind.InnerExpression,
+		expression,
+		flags: 0,
+		pos: startPos,
+		endPos,
+	};
+	return statement;
+}
+
+export function createPrefixToken(value: PrefixToken["value"]): PrefixToken {
+	return { kind: CmdSyntaxKind.PrefixToken, value, flags: 0 };
+}
+
+export function createPrefixExpression(
+	prefix: PrefixExpression["prefix"],
+	expression: PrefixExpression["expression"],
+): PrefixExpression {
+	return { kind: CmdSyntaxKind.PrefixExpression, prefix, expression, flags: 0 };
+}
+
+export function createCommandSource(children: CommandSource["children"]) {
+	const statement: CommandSource = { kind: CmdSyntaxKind.Source, children, flags: 0 };
+	for (const child of statement.children) {
+		child.parent = statement;
+	}
+	return statement;
+}
+
+export function createStringNode(text: string, quotes?: string): StringLiteral {
+	return { kind: CmdSyntaxKind.String, text, quotes, flags: 0 };
+}
+
+export function createNumberNode(value: number): NumberLiteral {
+	return { kind: CmdSyntaxKind.Number, value, flags: 0 };
+}
+
+export function createCommandName(name: StringLiteral): CommandName {
+	return {
+		kind: CmdSyntaxKind.CommandName,
+		name,
+		flags: 0,
+		pos: name.pos,
+		endPos: name.endPos,
+		rawText: name.rawText,
+	};
+}
+
+export function createIdentifier(name: string): Identifier {
+	return { kind: CmdSyntaxKind.Identifier, name, flags: 0 };
+}
+
+export function createOption(flag: string): Option {
+	return { kind: CmdSyntaxKind.Option, flag, flags: 0 };
+}
+
+export function createOperator(operator: OperatorToken["operator"]): OperatorToken {
+	return { kind: CmdSyntaxKind.OperatorToken, operator, flags: 0 };
+}
+
+export function createVariableDeclaration(
+	identifier: Identifier,
+	expression: VariableDeclaration["expression"],
+): VariableDeclaration {
+	return { kind: CmdSyntaxKind.VariableDeclaration, identifier, expression, flags: 0 };
+}
+
+export function createVariableStatement(declaration: VariableDeclaration): VariableStatement {
+	return { kind: CmdSyntaxKind.VariableStatement, declaration, flags: 0 };
+}
+
+export function createBooleanNode(value: boolean): BooleanLiteral {
+	return { kind: CmdSyntaxKind.Boolean, value, flags: 0 };
+}
+
+export function createEndOfStatementNode(): EndOfStatement {
+	return { kind: CmdSyntaxKind.EndOfStatement, flags: 0 };
+}
+
+export function createInvalidNode(
+	message: InvalidNode["message"],
+	nodes: Node[],
+	startPos?: number,
+	endPos?: number,
+): InvalidNode {
+	const firstNode = nodes[0];
+	const lastNode = nodes[nodes.size() - 1];
+
+	return {
+		kind: CmdSyntaxKind.Invalid,
+		message,
+		flags: NodeFlag.NodeHasError,
+		// eslint-disable-next-line roblox-ts/lua-truthiness
+		pos: startPos ?? firstNode.pos,
+		// eslint-disable-next-line roblox-ts/lua-truthiness
+		endPos: endPos ?? lastNode.endPos,
+	};
+}
+
+export function createBinaryExpression(
+	left: Node,
+	op: OperatorToken,
+	right: Node,
+	startPos?: number,
+	endPos?: number,
+): BinaryExpression {
+	const expression: BinaryExpression = {
+		kind: CmdSyntaxKind.BinaryExpression,
+		left,
+		operator: op,
+		right,
+		children: [left, right],
+		flags: 0,
+		pos: startPos,
+		endPos,
+	};
+	left.parent = expression;
+	right.parent = expression;
+	return expression;
+}
